# Conflicts:
#	cloud-modules/ideal-cloud-uua/ideal-cloud-uua.gradle
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/controllers/AuthorizationController.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/Application.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/Authorization.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/AuthorizationRecord.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/Department.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/GlobalLoginRestriction.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/LoginConfig.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/LoginRecord.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/Menu.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/RestrictionRule.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/Role.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/RoleMenu.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/User.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/UserLoginRestriction.java
#	cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/domain/entity/UserRole.java
#	spring-modules/spring-boot-autoconfigure/src/main/java/com/ziyao/ideal/boot/autoconfigure/web/mvc/WebMvcAutoConfiguration.java
#	spring-modules/spring-web/src/main/java/com/ziyao/ideal/web/base/BaseEntity.java

+++ cloud-modules/ideal-cloud-uua/src/main/java/com/ziyao/ideal/uua/repository/jpa/UserRepositoryJpa.java
@@ -10,7 +10,7 @@ import java.util.Optional;
  * @author ziyao
  */
 @Repository
-public interface UserRepositoryJpa extends JpaRepository<User, Long> {
+public interface UserRepositoryJpa extends JpaRepository<User, Integer> {
 
     Optional<User> findByUsername(String username);
